requirements.txt
confluent-kafka
requests
chess
python-dotenv
psycopg2-binary
